Changed LEDs to be 8x8

+++ src/lights.py
@@ -11,7 +11,7 @@ def updateLEDs(grid, x, y, sleep, period):
             if (y):
                 GPIO.output(y_channel_list[numy], GPIO.HIGH)
             if(sleep):
-                sleep(double(period))
+                sleep(float(period))
             GPIO.output(y_channel_list[numy], GPIO.HIGH)
             numy += 1
         numy =0
